+++ packages/generator-adp/test/unit/questions/helper/additional-messages.test.ts
@@ -1,6 +1,6 @@
 import { Severity } from '@sap-devx/yeoman-ui-types';
 
-import type { SourceApplication } from '@sap-ux/adp-tooling';
+import type { FlexUICapability, SourceApplication } from '@sap-ux/adp-tooling';
 import { AdaptationProjectType } from '@sap-ux/axios-extension';
 
 import {
@@ -18,12 +18,12 @@ describe('additional-messages', () => {
 
     describe('getSystemAdditionalMessages', () => {
         it('should return undefined if flexUISystem is undefined', () => {
-            const result = getSystemAdditionalMessages(undefined, false);
+            const result = getSystemAdditionalMessages(undefined, undefined);
             expect(result).toBeUndefined();
         });
 
         it('should return CLOUD_READY info message for cloud project', () => {
-            const result = getSystemAdditionalMessages(undefined, true);
+            const result = getSystemAdditionalMessages({} as FlexUICapability, AdaptationProjectType.CLOUD_READY);
             expect(result).toEqual({
                 message: `${t('prompts.projectTypeLabel')}: ${AdaptationProjectType.CLOUD_READY}`,
                 severity: Severity.information
@@ -31,7 +31,10 @@ describe('additional-messages', () => {
         });
 
         it('should return not deployable and not flex enabled error if not on-premise and not UIFlex', () => {
-            const result = getSystemAdditionalMessages({ isOnPremise: false, isUIFlex: false }, false);
+            const result = getSystemAdditionalMessages(
+                { isDtaFolderDeploymentSupported: false, isUIFlexSupported: false },
+                AdaptationProjectType.ON_PREMISE
+            );
             expect(result).toEqual({
                 message: t('error.notDeployableNotFlexEnabledSystemError'),
                 severity: Severity.warning
@@ -39,7 +42,10 @@ describe('additional-messages', () => {
         });
 
         it('should return not deployable system error if not on-premise but UIFlex is true', () => {
-            const result = getSystemAdditionalMessages({ isOnPremise: false, isUIFlex: true }, false);
+            const result = getSystemAdditionalMessages(
+                { isDtaFolderDeploymentSupported: false, isUIFlexSupported: true },
+                AdaptationProjectType.ON_PREMISE
+            );
             expect(result).toEqual({
                 message: t('error.notDeployableSystemError'),
                 severity: Severity.error
@@ -47,7 +53,10 @@ describe('additional-messages', () => {
         });
 
         it('should return not flex enabled warning if on-premise but UIFlex is false', () => {
-            const result = getSystemAdditionalMessages({ isOnPremise: true, isUIFlex: false }, false);
+            const result = getSystemAdditionalMessages(
+                { isDtaFolderDeploymentSupported: true, isUIFlexSupported: false },
+                AdaptationProjectType.ON_PREMISE
+            );
             expect(result).toEqual({
                 message: t('error.notFlexEnabledError'),
                 severity: Severity.warning
@@ -55,7 +64,10 @@ describe('additional-messages', () => {
         });
 
         it('should return ON_PREMISE info if on-premise and UIFlex is true', () => {
-            const result = getSystemAdditionalMessages({ isOnPremise: true, isUIFlex: true }, false);
+            const result = getSystemAdditionalMessages(
+                { isDtaFolderDeploymentSupported: true, isUIFlexSupported: true },
+                AdaptationProjectType.ON_PREMISE
+            );
             expect(result).toEqual({
                 message: `${t('prompts.projectTypeLabel')}: ${AdaptationProjectType.ON_PREMISE}`,
                 severity: Severity.information