+++ packages/ui5-test-writer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/ui5-test-writer
 
+## 0.7.69
+
+### Patch Changes
+
+- @sap-ux/project-access@1.35.4
+- @sap-ux/ui5-application-writer@1.7.9
+
 ## 0.7.68
 
 ### Patch Changes